LayoutTests/platform/gtk/imported/w3c/web-platform-tests/mathml/presentation-markup/operators/operator-dictionary-combining-expected.txt


PASS Spacing around ≠
PASS Spacing around =⃒
FAIL Spacing around |̸ assert_approx_equals: expected 61.09375 +/- 1 but got 77.75
FAIL Spacing around |⃒ assert_approx_equals: expected 61.09375 +/- 1 but got 77.75
FAIL Spacing around ⋉̸ assert_approx_equals: expected 72.21875 +/- 1 but got 77.75
FAIL Spacing around ⋉⃒ assert_approx_equals: expected 72.21875 +/- 1 but got 77.75
FAIL Spacing around ∄ assert_approx_equals: expected 58.3125 +/- 1 but got 77.75
FAIL Spacing around ∃⃒ assert_approx_equals: expected 58.3125 +/- 1 but got 77.75
 
=
 

 
≠
 

 
=⃒
 

 
|
 

 
|̸
 

 
|⃒
 

 
⋉
 

 
⋉̸
 

 
⋉⃒
 

 
∃
 

 
∄
 

 
∃⃒
 
